+++ src/index.ts
@@ -9,8 +9,8 @@ interface API {
 export async function activate(context: ExtensionContext): Promise<API> {
   let { subscriptions } = context
   const pluginManager = new PluginManager()
-  const service = new TsserverService(pluginManager, context.subscriptions, context)
-  subscriptions.push(services.regist(service))
+  const service = new TsserverService(pluginManager, context)
+  subscriptions.push(services.register(service))
 
   return {
     configurePlugin: (pluginId: string, configuration: {}): void => {