+++ history.md
@@ -30,6 +30,7 @@
   by default, use `tsserver.useLocalTsdk` to enable tsserver from workspace
   folder.
 - `watchBuild` use resolved config file returned from tsserver by default.
+- Add command `tsserver.chooseVersion`
 
 # 1.12.0